Fix javascript and remove resourceOptions
