+++ src/OrchardCore.Modules/OrchardCore.Forms/wwwroot/Scripts/formsWidgetEditor.js
@@ -1,0 +1,68 @@
+/*
+** NOTE: This file is generated by Gulp and should not be edited directly!
+** Any changes made directly to this file will be overwritten next time its asset group is processed by Gulp.
+*/
+
+formsWidgetEditor = function () {
+  var initilizeFieldType = function initilizeFieldType(wrapper) {
+    var selectMenus = wrapper.getElementsByClassName('field-type-select-menu');
+
+    for (var x = 0; x < selectMenus.length; x++) {
+      selectMenus[x].addEventListener('change', function (e) {
+        var labelTextContainer = wrapper.querySelector('.label-text-container');
+        var labelOption = wrapper.querySelector('.field-label-option-select-menu');
+        var visibleForInputContainers = wrapper.getElementsByClassName('show-for-input');
+
+        if (e.target.value == 'reset' || e.target.value == 'submit' || e.target.value == 'hidden') {
+          for (var i = 0; i < visibleForInputContainers.length; i++) {
+            visibleForInputContainers[i].classList.add('d-none');
+          }
+
+          labelTextContainer.classList.add('d-none');
+        } else {
+          for (var _i = 0; _i < visibleForInputContainers.length; _i++) {
+            visibleForInputContainers[_i].classList.remove('d-none');
+          }
+
+          labelOption.dispatchEvent(new Event('change'));
+        }
+      });
+    }
+  };
+
+  var initilizeLabelOptions = function initilizeLabelOptions(wrapper) {
+    var labelOptions = wrapper.getElementsByClassName('field-label-option-select-menu');
+
+    var _loop = function _loop(x) {
+      labelOptions[x].addEventListener('change', function (e) {
+        var container = labelOptions[x].closest('.properties-wrapper');
+
+        if (container == null) {
+          contine;
+        }
+
+        var labelContainer = container.querySelector('.label-text-container');
+
+        if (e.target.value != 'None') {
+          labelContainer.classList.remove('d-none');
+        } else {
+          labelContainer.classList.add('d-none');
+        }
+      });
+    };
+
+    for (var x = 0; x < labelOptions.length; x++) {
+      _loop(x);
+    }
+  };
+
+  return {
+    initilizeFieldType: initilizeFieldType,
+    initilizeLabelOptions: initilizeLabelOptions
+  };
+}();
+
+document.addEventListener('DOMContentLoaded', function () {
+  formsWidgetEditor.initilizeFieldType(document);
+  formsWidgetEditor.initilizeLabelOptions(document);
+});
